Moved tests to existing file
